feat(util): 사용자 입력을 숫자로 변환시 `Long` 타입으로 변환하도록 수정

+++ src/main/java/lotto/util/NumberArrayParser.java
@@ -10,7 +10,7 @@ public class NumberArrayParser {
     private NumberArrayParser() {
     }
 
-    public static List<Integer> parse(String input) {
+    public static List<Long> parse(String input) {
         if (input.isEmpty()) {
             return new ArrayList<>();
         }
